Move parsing tests to parse module.
